# api/openapi.yaml
openapi: 3.0.1
info:
  title: VMClarity APIs
  version: 1.0.0
paths:
...
  /findings:
    get:
      summary: Get all findings.
      operationId: GetFindings
      parameters:
        - $ref: '#/components/parameters/odataFilter'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataCount'
        - $ref: '#/components/parameters/odataTop'
        - $ref: '#/components/parameters/odataSkip'
        - $ref: '#/components/parameters/odataExpand'
        - $ref: '#/components/parameters/odataOrderBy'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Findings'
        default:
          $ref: '#/components/responses/UnknownError'
    post:
      summary: Create a finding
      operationId: PostFindings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Finding'
        required: true
      responses:
        201:
          description: A new finding was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        400:
          description: Invalid finding supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        409:
          description: Finding conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingExists'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
  /findings/{findingID}:
    get:
      summary: Get the details for a finding.
      operationId: GetFindingsFindingID
      parameters:
        - $ref: '#/components/parameters/findingID'
        - $ref: '#/components/parameters/odataSelect'
        - $ref: '#/components/parameters/odataExpand'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        404:
          description: Finding ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
    put:
      summary: Update a finding.
      operationId: PutFindingsFindingID
      parameters:
        - $ref: '#/components/parameters/findingID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Finding'
        required: true
      responses:
        200:
          description: Updated finding successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        400:
          description: Invalid finding supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Finding ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    patch:
      summary: Patch a finding.
      operationId: PatchFindingsFindingID
      parameters:
        - $ref: '#/components/parameters/findingID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Finding'
        required: true
      responses:
        200:
          description: Patched finding successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Finding'
        400:
          description: Invalid finding supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        404:
          description: Finding ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
      x-codegen-request-body-name: body
    delete:
      summary: Delete a finding.
      operationId: DeleteFindingsFindingID
      parameters:
        - $ref: '#/components/parameters/findingID'
      responses:
        200:
          $ref: '#/components/responses/Success'
        404:
          description: Finding ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          $ref: '#/components/responses/UnknownError'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in all cases of failures.
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          readOnly: true
      description: An object that is returned in cases of success that returns nothing.
    Metadata:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
    MetadataReadOnly:
      type: object
      properties:
        annotations:
          $ref: '#/components/schemas/Annotations'
          readOnly: true
    Annotations:
      type: array
      description: Generic map of string keys and string values to attach arbitrary non-identifying metadata to objects.
      items:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
...
    ScanRelationship:
      type: object
      description: Describes an expandable relationship to Scan object
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            name:
              type: string
              readOnly: true
            revision:
              type: integer
              readOnly: true
            startTime:
              type: string
              format: date-time
              readOnly: true
            endTime:
              type: string
              format: date-time
              readOnly: true
            scanConfig:
              $ref: '#/components/schemas/ScanConfigRelationship'
              readOnly: true
            scope:
              description: |
                The query used to limit the scope of this scan. It uses
                the ODATA $filter query language to limit the collection of assets
                that this scan will operate over. For example
                `startswith(assetInfo.location, 'eu-west-2')` will limit this scan
                to just assets in the eu-west-2 AWS region.
              type: string
            assetScanTemplate:
              $ref: '#/components/schemas/AssetScanTemplate'
              readOnly: true
            assetIDs:
              description: List of asset IDs to be scanned
              type: array
              items:
                type: string
              nullable: true
              readOnly: true
            status:
              $ref: '#/components/schemas/ScanStatus'
              readOnly: true
            summary:
              $ref: '#/components/schemas/ScanSummary'
              readOnly: true
          required:
            - id
        - $ref: '#/components/schemas/ScanTemplateReadOnly'
...
    ScanFindingsSummary:
      description: A summary of the scan findings.
      type: object
      properties:
        totalPackages:
          type: integer
        totalExploits:
          type: integer
        totalMalware:
          type: integer
        totalMisconfigurations:
          type: integer
        totalRootkits:
          type: integer
        totalSecrets:
          type: integer
        totalInfoFinder:
          type: integer
        totalVulnerabilities:
          $ref: '#/components/schemas/VulnerabilityScanSummary'
...
    AssetRelationship:
      type: object
      description: Describes a relationship to an asset which can be expanded.
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            revision:
              type: integer
              readOnly: true
            assetInfo:
              $ref: '#/components/schemas/AssetType'
              readOnly: true
            scansCount:
              description: Total number of scans that have ever run for this asset
              type: integer
              readOnly: true
            firstSeen:
              type: string
              format: date-time
            terminatedOn:
              type: string
              format: date-time
              readOnly: true
            lastSeen:
              type: string
              format: date-time
              readOnly: true
            summary:
              $ref: '#/components/schemas/ScanFindingsSummary'
              readOnly: true
            providers:
              type: array
              description: List of providers which discover the asset.
              items:
                $ref: '#/components/schemas/ProviderRelationship'
              readOnly: true
          required: ['id']
...
    AssetType:
      type: object
      anyOf:
        - $ref: '#/components/schemas/VMInfo'
        - $ref: '#/components/schemas/PodInfo'
        - $ref: '#/components/schemas/DirInfo'
        - $ref: '#/components/schemas/ContainerImageInfo'
        - $ref: '#/components/schemas/ContainerInfo'
      discriminator:
        propertyName: objectType
        mapping:
          VMInfo: '#/components/schemas/VMInfo'
          PodInfo: '#/components/schemas/PodInfo'
          DirInfo: '#/components/schemas/DirInfo'
          ContainerImageInfo: '#/components/schemas/ContainerImageInfo'
          ContainerInfo: '#/components/schemas/ContainerInfo'
...
    ResourceCleanupStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the state of resource cleanup.
            
            | State   | Description                                                |
            | ------- | ---------------------------------------------------------- |
            | Pending | Initial state for cleaning up resources                    |
            | Skipped | Resource cleanup has been skipped due to Delete Job Policy |
            | Failed  | Cleaning up resources has been failed                      |
            | Done    | Resources have been successfully cleaned up                |
          type: string
          enum:
            - Pending
            - Skipped
            - Failed
            - Done
        reason:
          description: |
            Machine readable reason for state transition.

            | State   | Reason          | Description                                                            |
            | ------- | --------------- | ---------------------------------------------------------------------- |
            | Pending | AssetScanCreate | AssetScan created                                                      |
            | Skipped | NotApplicable   | Resource cleanup is not required as it is managed outside of VMClarity |
            | Skipped | DeletePolicy    | Resource cleanup has been skipped due to Delete Job Policy             |
            | Failed  | ProviderError   | Failed due to Provider error                                           |
            | Failed  | InternalError   | Failed due to internal error                                           |
            | Done    | Success         | Successfully completed                                                 |
          type: string
          enum:
            - AssetScanCreated
            - NotApplicable
            - DeletePolicy
            - ProviderError
            - InternalError
            - Success
        message:
          description: Human readable message.
          type: string
        lastTransitionTime:
          description: Last date time when the status has changed.
          type: string
          format: date-time
...
    AssetScanStatus:
      type: object
      required:
        - state
        - reason
        - lastTransitionTime
      properties:
        state:
          description: |
            Describes the state of scan result.

            | State       | Description                                                                                      |
            | ----------- | ------------------------------------------------------------------------------------------------ |
            | Pending     | Initial state for AssetScan waiting for being scheduled                                          |
            | Scheduled   | AssetScan which has been scheduled on Provider                                                   |
            | ReadyToScan | Provider acknowledged that scanners for AssetScan is ready to run                                |
            | InProgress  | Scanners are running on the Target                                                               |
            | Aborted     | AssetScan has been aborted and all running Scanners need to be cancelled and shutdown gracefully |
            | Failed      | Running Scanners on Target has failed, check *reason* and *message* fields for the details       |
            | Done        | Running Scanners on Target has finished with no errors                                           |
          type: string
          enum:
            - Pending
            - Scheduled
            - ReadyToScan
            - InProgress
            - Aborted
            - Failed
            - Done
        reason:
          description: |
            Machine readable reason for state transition.

            | State       | Reason           | Description                                                                                |
            | ----------- | ---------------- | ------------------------------------------------------------------------------------------ |
            | Pending     | Created          | Scan is pending as it has been newly created                                               |
            | Scheduled   | Provisioning     | Scan is scheduled as Asset resources are being provisioned                                 |
            | ReadyToScan | UnSupervised     | Scan is created manually without the involvement of the Orchestrator (e.g. CI/CD scenario) |
            | ReadyToScan | ResourcesReady   | Scan is ready to scan as Asset resources have been provisioned                             |
            | InProgress  | ScannerIsRunning | Scan is in progress as Scanner is still performing scanning on Asset resources             |
            | Aborted     | Cancellation     | Scan has been aborted due to cancellation request                                          |
            | Aborted     | Timeout          | Scan has been aborted due to timeout                                                       |
            | Failed      | Error            | Scan has failed to one or more errors reported by the Scanner                              |
            | Failed      | AbortTimeout     | Scan has failed due to being in Aborted state for too long                                 |
            | Done        | Success          | Scan has finished successfully                                                             |
          type: string
          enum:
            - Created
            - Provisioning
            - UnSupervised
            - ResourcesReady
            - ScannerIsRunning
            - Cancellation
            - Timeout
            - Error
            - AbortTimedOut
            - Success
          x-enum-varnames:
            - AssetScanStatusReasonCreated
            - AssetScanStatusReasonProvisioning
            - AssetScanStatusReasonUnSupervised
            - AssetScanStatusReasonResourcesReady
            - AssetScanStatusReasonScannerIsRunning
            - AssetScanStatusReasonCancellation
            - AssetScanStatusReasonTimeout
            - AssetScanStatusReasonError
            - AssetScanStatusReasonAbortTimeout
            - AssetScanStatusReasonSuccess
        message:
          description: Human readable message.
          type: string
        lastTransitionTime:
          description: Last date time when the status has changed.
          type: string
          format: date-time
...
    AssetScanStats:
      type: object
      properties:
        general:
          $ref: '#/components/schemas/AssetScanGeneralStats'
        sbom:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        malware:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        rootkits:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        misconfigurations:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        exploits:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
        infoFinder:
          type: array
          items:
            $ref: '#/components/schemas/AssetScanInputScanStats'
...
    AssetScanRelationship:
      type: object
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
              # TODO(sambetts) Decide if we want the validation here by having
              # separate schemas for GET, POST and PATCH.
              #
              # We can't make anything readOnly or required here if we want to
              # use the same schema for PATCH otherwise it will require you
              # always send the required fields and remove the readOnly fields
              # on the client side.
              #
              # readOnly: true
            revision:
              type: integer
              readOnly: true
            asset:
              $ref: '#/components/schemas/AssetRelationship'
              readOnly: true
            scan:
              $ref: '#/components/schemas/ScanRelationship'
              readOnly: true
            status:
              $ref: '#/components/schemas/AssetScanStatus'
              readOnly: true
            sbom:
              $ref: '#/components/schemas/SbomScan'
              readOnly: true
            vulnerabilities:
              $ref: '#/components/schemas/VulnerabilityScan'
              readOnly: true
            malware:
              $ref: '#/components/schemas/MalwareScan'
              readOnly: true
            rootkits:
              $ref: '#/components/schemas/RootkitScan'
              readOnly: true
            secrets:
              $ref: '#/components/schemas/SecretScan'
              readOnly: true
            misconfigurations:
              $ref: '#/components/schemas/MisconfigurationScan'
              readOnly: true
            exploits:
              $ref: '#/components/schemas/ExploitScan'
              readOnly: true
            infoFinder:
              $ref: '#/components/schemas/InfoFinderScan'
              readOnly: true
            findingsProcessed:
              type: boolean
              readOnly: true
            resourceCleanupStatus:
              $ref: '#/components/schemas/ResourceCleanupStatus'
              readOnly: true
            stats:
              $ref: '#/components/schemas/AssetScanStats'
              readOnly: true
            summary:
              $ref: '#/components/schemas/ScanFindingsSummary'
              readOnly: true
          required: [id]
        - $ref: '#/components/schemas/AssetScanTemplateReadOnly'
...
    ProviderRelationship:
      type: object
      description: Describes a relationship to a provider which can be expanded.
      allOf:
        - $ref: '#/components/schemas/MetadataReadOnly'
        - type: object
          properties:
            id:
              type: string
            revision:
              type: integer
              readOnly: true
            displayName:
              description: Human-readable name for the provider.
              type: string
              readOnly: true
            lastHeartbeatTime:
              description: Timestamp of the last heartbeat from the provider.
              type: string
              format: date-time
              readOnly: true
            status:
              $ref: '#/components/schemas/ProviderStatus'
              readOnly: true
            providerRuntimeVersion:
              description: Version of the provider runtime that the provider was compiled with.
              type: string
              readOnly: true
...
    SbomScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        packages:
          type: array
          items:
            $ref: '#/components/schemas/Package'
          nullable: true
    VulnerabilityScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
          nullable: true
    MalwareScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        malware:
          type: array
          items:
            $ref: '#/components/schemas/Malware'
          nullable: true
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/ScannerMetadata'
          nullable: true
    RootkitScan:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ScannerStatus'
        rootkits:
          type: array
          items:
            $ref: '#/components/schemas/Rootkit'
          nullable: true
    SecretScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        secrets:
          type: array
          items:
            $ref: '#/components/schemas/Secret'
          nullable: true
    MisconfigurationScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        scanners:
          type: array
          items:
            type: string
          nullable: true
        misconfigurations:
          type: array
          items:
            $ref: '#/components/schemas/Misconfiguration'
          nullable: true
    InfoFinderScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        scanners:
          type: array
          items:
            type: string
          nullable: true
        infos:
          type: array
          items:
            $ref: '#/components/schemas/InfoFinderInfo'
          nullable: true
...
    ExploitScan:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ScannerStatus"
        exploits:
          type: array
          items:
            $ref: '#/components/schemas/Exploit'
          nullable: true
...
    FindingExists:
      type: object
      properties:
        message:
          description: Describes which unique constraint combination causes the conflict.
          type: string
          readOnly: true
        finding:
          $ref: '#/components/schemas/Finding'
    Findings:
      type: object
      properties:
        count:
          description: Total findings count according to the given filters
          type: integer
        items:
          description: List of findings according to the given filters
          type: array
          items:
            $ref: '#/components/schemas/Finding'
    PackageFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Package'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    VulnerabilityFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Vulnerability'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    MalwareFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Malware'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    SecretFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Secret'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    MisconfigurationFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Misconfiguration'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    RootkitFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Rootkit'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    ExploitFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/Exploit'
        - type: object
          properties:
            objectType:
              type: string
          required: [objectType]
    InfoFinderFindingInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/InfoFinderInfo'
        - type: object
          properties:
            objectType:
              type: string
          required: [ objectType ]
    Finding:
      type: object
      allOf:
        - $ref: '#/components/schemas/Metadata'
        - type: object
          properties:
            id:
              type: string
              # TODO(sambetts) Decide if we want the validation here by having
              # separate schemas for GET, POST and PATCH
              #
              # We can't make anything readOnly or required here if we want to
              # use the same schema for PATCH otherwise it will require you
              # always send the required fields and remove the readOnly fields
              # on the client side.
              #
              # readOnly: true
            asset:
              $ref: '#/components/schemas/AssetRelationship'
            foundBy:
              $ref: '#/components/schemas/AssetScanRelationship'
            foundOn:
              description: When this finding was discovered by a scan
              type: string
              format: date-time
            invalidatedOn:
              description: When this finding was invalidated by a newer scan
              type: string
              format: date-time
            findingInfo:
              anyOf:
                - $ref: '#/components/schemas/PackageFindingInfo'
                - $ref: '#/components/schemas/VulnerabilityFindingInfo'
                - $ref: '#/components/schemas/MalwareFindingInfo'
                - $ref: '#/components/schemas/SecretFindingInfo'
                - $ref: '#/components/schemas/MisconfigurationFindingInfo'
                - $ref: '#/components/schemas/RootkitFindingInfo'
                - $ref: '#/components/schemas/ExploitFindingInfo'
                - $ref: '#/components/schemas/InfoFinderFindingInfo'
              discriminator:
                propertyName: objectType
                mapping:
                  Package: '#/components/schemas/PackageFindingInfo'
                  Vulnerability: '#/components/schemas/VulnerabilityFindingInfo'
                  Malware: '#/components/schemas/MalwareFindingInfo'
                  Secret: '#/components/schemas/SecretFindingInfo'
                  Misconfiguration: '#/components/schemas/MisconfigurationFindingInfo'
                  Rootkit: '#/components/schemas/RootkitFindingInfo'
                  Exploit: '#/components/schemas/ExploitFindingInfo'
                  InfoFinder: '#/components/schemas/InfoFinderFindingInfo'
...
  responses:
    Success:
      description: Success message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResponse'
    UnknownError:
      description: Unknown error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  parameters:
...
    odataFilter:
      name: "$filter"
      in: query
      schema:
        type: string
    odataSelect:
      name: "$select"
      in: query
      schema:
        type: string
    odataCount:
      name: "$count"
      in: query
      schema:
        type: boolean
    odataTop:
      name: "$top"
      in: query
      schema:
        type: integer
    odataSkip:
      name: "$skip"
      in: query
      schema:
        type: integer
    odataExpand:
      name: "$expand"
      in: query
      schema:
        type: string
    odataOrderBy:
      name: "$orderby"
      in: query
      schema:
        type: string
      x-go-name: "OrderBy"
...
    findingID:
      name: findingID
      in: path
      required: true
      schema:
        type: string
...